Fix mbedtls (#6)
The .bbappend was fixed to version 2.28.0. meta-openembedded
has upgraded mbedtls in 60e8a5e23a82838b731812837ec1d170fb9ddccb
to use version 2.28.2 which caused a problem that the .bbappend
cannot find the default recipe. This is fixed by removing
the reference to a fixed version in the .bbappend name.

Signed-off-by: Mohannad Oraby <mohannad.oraby@chargebyte.com>